Move toolbar to its own file.
